Add files via upload
Product classification is the process of classifying service products. In this process, the products are divided according to specific characteristics or in other words, it is the process of grouping the products into different categories based on their properties and relationships to other products. There is no such formal classification system exists and even for the identification of the products every industry has their own standardised methods of product classification. Product classification can be done in a variety of perspectives and it depends on the consumer and the business products. Product classification contains a huge number of categories and millions of products and each product must be mapped to the corresponding category. Hence this requires a tremendous effort by integrating different predictive analytical techniques including data mining, statistical modelling and machine learning which helps the analysts to build a predictive model which would be able to distinguish